[monarch] refactor monarch/ callsites to use monarch.actor
as title

Differential Revision: [D77823401](https://our.internmc.facebook.com/intern/diff/D77823401/)

[ghstack-poisoned]
